+++ Entities/LimbPath.cpp
@@ -563,7 +563,6 @@ bool LimbPath::RestartFree(Vector &limbPos, MOID MOIDToIgnore, int ignoreTeam)
     float result = 0;
 
     g_SceneMan.GetTerrain()->LockBitmaps();
-    acquire_bitmap(g_SceneMan.GetMOIDBitmap());
     
     if (IsStaticPoint())
 	{
@@ -639,7 +638,7 @@ bool LimbPath::RestartFree(Vector &limbPos, MOID MOIDToIgnore, int ignoreTeam)
             found = true;
         }
     }
-    release_bitmap(g_SceneMan.GetMOIDBitmap());
+    
     g_SceneMan.GetTerrain()->UnlockBitmaps();
 
     if (found)