+++ Managers/SettingsMan.cpp
@@ -44,9 +44,7 @@ namespace RTE {
 		m_ShowForeignItems = true;
 		m_ShowMetaScenes = false;
 
-		m_DisableLuaJIT = false;
-		m_RecommendedMOIDCount = 512;
-		m_SimplifiedCollisionDetection = false;
+		m_RecommendedMOIDCount = 240;
 		m_SceneBackgroundAutoScaleMode = 1;
 		m_DisableFactionBuyMenuThemes = false;
 		m_DisableFactionBuyMenuThemeCursors = false;
@@ -137,85 +135,155 @@ namespace RTE {
 			reader >> g_AudioMan.m_MinimumDistanceForPanning;
 		//////////////////////////////////////////////////
 
-		});
-		MatchProperty("ShowForeignItems", { reader >> m_ShowForeignItems; });
-		MatchProperty("FlashOnBrainDamage", { reader >> m_FlashOnBrainDamage; });
-		MatchProperty("BlipOnRevealUnseen", { reader >> m_BlipOnRevealUnseen; });
-		MatchProperty("MaxUnheldItems", { reader >> g_MovableMan.m_MaxDroppedItems; });
-		MatchProperty("UnheldItemsHUDDisplayRange", { SetUnheldItemsHUDDisplayRange(std::stof(reader.ReadPropValue())); });
-		MatchProperty("AlwaysDisplayUnheldItemsInStrategicMode", { reader >> m_AlwaysDisplayUnheldItemsInStrategicMode; });
-		MatchProperty("SubPieMenuHoverOpenDelay", { reader >> m_SubPieMenuHoverOpenDelay; });
-		MatchProperty("EndlessMode", { reader >> m_EndlessMetaGameMode; });
-		MatchProperty("EnableCrabBombs", { reader >> m_EnableCrabBombs; });
-		MatchProperty("CrabBombThreshold", { reader >> m_CrabBombThreshold; });
-		MatchProperty("ShowEnemyHUD", { reader >> m_ShowEnemyHUD; });
-		MatchProperty("SmartBuyMenuNavigation", { reader >> m_EnableSmartBuyMenuNavigation; });
-		MatchProperty("ScrapCompactingHeight", { reader >> g_SceneMan.m_ScrapCompactingHeight; });
-		MatchProperty("AutomaticGoldDeposit", { reader >> m_AutomaticGoldDeposit; });
-		MatchProperty("ScreenShakeStrength", { reader >> g_CameraMan.m_ScreenShakeStrength; });
-		MatchProperty("ScreenShakeDecay", { reader >> g_CameraMan.m_ScreenShakeDecay; });
-		MatchProperty("MaxScreenShakeTime", { reader >> g_CameraMan.m_MaxScreenShakeTime; });
-		MatchProperty("DefaultShakePerUnitOfGibEnergy", { reader >> g_CameraMan.m_DefaultShakePerUnitOfGibEnergy; });
-		MatchProperty("DefaultShakePerUnitOfRecoilEnergy", { reader >> g_CameraMan.m_DefaultShakePerUnitOfRecoilEnergy; });
-		MatchProperty("DefaultShakeFromRecoilMaximum", { reader >> g_CameraMan.m_DefaultShakeFromRecoilMaximum; });
-		MatchProperty("LaunchIntoActivity", { reader >> g_ActivityMan.m_LaunchIntoActivity; });
-		MatchProperty("DefaultActivityType", { reader >> g_ActivityMan.m_DefaultActivityType; });
-		MatchProperty("DefaultActivityName", { reader >> g_ActivityMan.m_DefaultActivityName; });
-		MatchProperty("DefaultSceneName", { reader >> g_SceneMan.m_DefaultSceneName; });
-		MatchProperty("DisableLuaJIT", { reader >> m_DisableLuaJIT; });
-		MatchProperty("RecommendedMOIDCount", { reader >> m_RecommendedMOIDCount; });
-		MatchProperty("SimplifiedCollisionDetection", { reader >> m_SimplifiedCollisionDetection; });
-		MatchProperty("SceneBackgroundAutoScaleMode", { SetSceneBackgroundAutoScaleMode(std::stoi(reader.ReadPropValue())); });
-		MatchProperty("DisableFactionBuyMenuThemes", { reader >> m_DisableFactionBuyMenuThemes; });
-		MatchProperty("DisableFactionBuyMenuThemeCursors", { reader >> m_DisableFactionBuyMenuThemeCursors; });
-		MatchProperty("PathFinderGridNodeSize", { reader >> m_PathFinderGridNodeSize; });
-		MatchProperty("AIUpdateInterval", { reader >> m_AIUpdateInterval; });
-		MatchProperty("EnableParticleSettling", { reader >> g_MovableMan.m_SettlingEnabled; });
-		MatchProperty("EnableMOSubtraction", { reader >> g_MovableMan.m_MOSubtractionEnabled; });
-		MatchProperty("DeltaTime", { g_TimerMan.SetDeltaTimeSecs(std::stof(reader.ReadPropValue())); });
-		MatchProperty("AllowSavingToBase", { reader >> m_AllowSavingToBase; });
-		MatchProperty("ShowMetaScenes", { reader >> m_ShowMetaScenes; });
-		MatchProperty("SkipIntro", { reader >> m_SkipIntro; });
-		MatchProperty("ShowToolTips", { reader >> m_ShowToolTips; });
-		MatchProperty("CaseSensitiveFilePaths", { System::EnableFilePathCaseSensitivity(std::stoi(reader.ReadPropValue())); });
-		MatchProperty("DisableLoadingScreenProgressReport", { reader >> m_DisableLoadingScreenProgressReport; });
-		MatchProperty("LoadingScreenProgressReportPrecision", { reader >> m_LoadingScreenProgressReportPrecision; });
-		MatchProperty("ConsoleScreenRatio", { g_ConsoleMan.SetConsoleScreenSize(std::stof(reader.ReadPropValue())); });
-		MatchProperty("ConsoleUseMonospaceFont", { reader >> g_ConsoleMan.m_ConsoleUseMonospaceFont; });
-		MatchProperty("AdvancedPerformanceStats", { reader >> g_PerformanceMan.m_AdvancedPerfStats; });
-		MatchProperty("MenuTransitionDurationMultiplier", { SetMenuTransitionDurationMultiplier(std::stof(reader.ReadPropValue())); });
-		MatchProperty("DrawAtomGroupVisualizations", { reader >> m_DrawAtomGroupVisualizations; });
-		MatchProperty("DrawHandAndFootGroupVisualizations", { reader >> m_DrawHandAndFootGroupVisualizations; });
-		MatchProperty("DrawLimbPathVisualizations", { reader >> m_DrawLimbPathVisualizations; });
-		MatchProperty("DrawRaycastVisualizations", { reader >> g_SceneMan.m_DrawRayCastVisualizations; });
-		MatchProperty("DrawPixelCheckVisualizations", { reader >> g_SceneMan.m_DrawPixelCheckVisualizations; });
-		MatchProperty("PrintDebugInfo", { reader >> m_PrintDebugInfo; });
-		MatchProperty("MeasureModuleLoadTime", { reader >> m_MeasureModuleLoadTime; });
-		MatchProperty("PlayerNetworkName", { reader >> m_PlayerNetworkName; });
-		MatchProperty("NetworkServerName", { reader >> m_NetworkServerAddress; });
-		MatchProperty("UseNATService", { reader >> g_NetworkServer.m_UseNATService; });
-		MatchProperty("NATServiceAddress", { reader >> m_NATServiceAddress; });
-		MatchProperty("NATServerName", { reader >> m_NATServerName; });
-		MatchProperty("NATServerPassword", { reader >> m_NATServerPassword; });
-		MatchProperty("UseExperimentalMultiplayerSpeedBoosts", { reader >> m_UseExperimentalMultiplayerSpeedBoosts; });
-		MatchProperty("ClientInputFps", { reader >> g_NetworkClient.m_ClientInputFps; });
-		MatchProperty("ServerTransmitAsBoxes", { reader >> g_NetworkServer.m_TransmitAsBoxes; });
-		MatchProperty("ServerBoxWidth", { reader >> g_NetworkServer.m_BoxWidth; });
-		MatchProperty("ServerBoxHeight", { reader >> g_NetworkServer.m_BoxHeight; });
-		MatchProperty("ServerUseHighCompression", { reader >> g_NetworkServer.m_UseHighCompression; });
-		MatchProperty("ServerUseFastCompression", { reader >> g_NetworkServer.m_UseFastCompression; });
-		MatchProperty("ServerUseDeltaCompression", { reader >> g_NetworkServer.m_UseDeltaCompression; });
-		MatchProperty("ServerHighCompressionLevel", { reader >> g_NetworkServer.m_HighCompressionLevel; });
-		MatchProperty("ServerFastAccelerationFactor", { reader >> g_NetworkServer.m_FastAccelerationFactor; });
-		MatchProperty("ServerUseInterlacing", { reader >> g_NetworkServer.m_UseInterlacing; });
-		MatchProperty("ServerEncodingFps", { reader >> g_NetworkServer.m_EncodingFps; });
-		MatchProperty("ServerSleepWhenIdle", { reader >> g_NetworkServer.m_SleepWhenIdle; });
-		MatchProperty("ServerSimSleepWhenIdle", { reader >> g_NetworkServer.m_SimSleepWhenIdle; });
-		MatchProperty("VisibleAssemblyGroup", { m_VisibleAssemblyGroupsList.push_back(reader.ReadPropValue()); });
-		MatchProperty("DisableMod", { m_DisabledMods.try_emplace(reader.ReadPropValue(), true); });
-		MatchProperty("EnableGlobalScript", { m_EnabledGlobalScripts.try_emplace(reader.ReadPropValue(), true); });
-		MatchProperty("MouseSensitivity", { reader >> g_UInputMan.m_MouseSensitivity; });
-		MatchForwards("Player1Scheme") MatchForwards("Player2Scheme") MatchForwards("Player3Scheme") MatchProperty("Player4Scheme", {
+		} else if (propName == "ShowForeignItems") {
+			reader >> m_ShowForeignItems;
+		} else if (propName == "FlashOnBrainDamage") {
+			reader >> m_FlashOnBrainDamage;
+		} else if (propName == "BlipOnRevealUnseen") {
+			reader >> m_BlipOnRevealUnseen;
+		} else if (propName == "MaxUnheldItems") {
+			reader >> g_MovableMan.m_MaxDroppedItems;
+		} else if (propName == "UnheldItemsHUDDisplayRange") {
+			SetUnheldItemsHUDDisplayRange(std::stof(reader.ReadPropValue()));
+		} else if (propName == "AlwaysDisplayUnheldItemsInStrategicMode") {
+			reader >> m_AlwaysDisplayUnheldItemsInStrategicMode;
+    } else if (propName == "SubPieMenuHoverOpenDelay") {
+			reader >> m_SubPieMenuHoverOpenDelay;
+		} else if (propName == "EndlessMode") {
+			reader >> m_EndlessMetaGameMode;
+		} else if (propName == "EnableCrabBombs") {
+			reader >> m_EnableCrabBombs;
+		} else if (propName == "CrabBombThreshold") {
+			reader >> m_CrabBombThreshold;
+		} else if (propName == "ShowEnemyHUD") {
+			reader >> m_ShowEnemyHUD;
+		} else if (propName == "SmartBuyMenuNavigation") {
+			reader >> m_EnableSmartBuyMenuNavigation;
+		} else if (propName == "ScreenShakeStrength") {
+			reader >> m_ScreenShakeStrength;
+		} else if (propName == "ScreenShakeDecay") {
+			reader >> m_ScreenShakeDecay;
+		} else if (propName == "MaxScreenShakeTime") {
+			reader >> m_MaxScreenShakeTime;
+		} else if (propName == "DefaultShakePerUnitOfGibEnergy") {
+			reader >> m_DefaultShakePerUnitOfGibEnergy;
+		} else if (propName == "DefaultShakePerUnitOfRecoilEnergy") {
+			reader >> m_DefaultShakePerUnitOfRecoilEnergy;
+		} else if (propName == "DefaultShakeFromRecoilMaximum") {
+			reader >> m_DefaultShakeFromRecoilMaximum;
+		} else if (propName == "LaunchIntoActivity") {
+			reader >> g_ActivityMan.m_LaunchIntoActivity;
+		} else if (propName == "DefaultActivityType") {
+			reader >> g_ActivityMan.m_DefaultActivityType;
+		} else if (propName == "DefaultActivityName") {
+			reader >> g_ActivityMan.m_DefaultActivityName;
+		} else if (propName == "DefaultSceneName") {
+			reader >> g_SceneMan.m_DefaultSceneName;
+		} else if (propName == "DisableLuaJIT") {
+			reader >> g_LuaMan.m_DisableLuaJIT;
+		} else if (propName == "RecommendedMOIDCount") {
+			reader >> m_RecommendedMOIDCount;
+		} else if (propName == "SceneBackgroundAutoScaleMode") {
+			SetSceneBackgroundAutoScaleMode(std::stoi(reader.ReadPropValue()));
+		} else if (propName == "DisableFactionBuyMenuThemes") {
+			reader >> m_DisableFactionBuyMenuThemes;
+		} else if (propName == "PathFinderGridNodeSize") {
+			reader >> m_PathFinderGridNodeSize;
+		} else if (propName == "AIUpdateInterval") {
+			reader >> m_AIUpdateInterval;
+		} else if (propName == "EnableParticleSettling") {
+			reader >> g_MovableMan.m_SettlingEnabled;
+		} else if (propName == "EnableMOSubtraction") {
+			reader >> g_MovableMan.m_MOSubtractionEnabled;
+		} else if (propName == "DeltaTime") {
+			g_TimerMan.SetDeltaTimeSecs(std::stof(reader.ReadPropValue()));
+		} else if (propName == "RealToSimCap") {
+			g_TimerMan.SetRealToSimCap(std::stof(reader.ReadPropValue()));
+		} else if (propName == "AllowSavingToBase") {
+			reader >> m_AllowSavingToBase;
+		} else if (propName == "ShowMetaScenes") {
+			reader >> m_ShowMetaScenes;
+		} else if (propName == "SkipIntro") {
+			reader >> m_SkipIntro;
+		} else if (propName == "ShowToolTips") {
+			reader >> m_ShowToolTips;
+		} else if (propName == "CaseSensitiveFilePaths") {
+			System::EnableFilePathCaseSensitivity(std::stoi(reader.ReadPropValue()));
+		} else if (propName == "DisableLoadingScreenProgressReport") {
+			reader >> m_DisableLoadingScreenProgressReport;
+		} else if (propName == "LoadingScreenProgressReportPrecision") {
+			reader >> m_LoadingScreenProgressReportPrecision;
+		} else if (propName == "ConsoleScreenRatio") {
+			g_ConsoleMan.SetConsoleScreenSize(std::stof(reader.ReadPropValue()));
+		} else if (propName == "ConsoleUseMonospaceFont") {
+			reader >> g_ConsoleMan.m_ConsoleUseMonospaceFont;
+		} else if (propName == "AdvancedPerformanceStats") {
+			reader >> g_PerformanceMan.m_AdvancedPerfStats;
+		} else if (propName == "MenuTransitionDurationMultiplier") {
+			SetMenuTransitionDurationMultiplier(std::stof(reader.ReadPropValue()));
+		} else if (propName == "DrawAtomGroupVisualizations") {
+			reader >> m_DrawAtomGroupVisualizations;
+		} else if (propName == "DrawHandAndFootGroupVisualizations") {
+			reader >> m_DrawHandAndFootGroupVisualizations;
+		} else if (propName == "DrawLimbPathVisualizations") {
+			reader >> m_DrawLimbPathVisualizations;
+		} else if (propName == "DrawRaycastVisualizations") {
+			reader >> g_SceneMan.m_DrawRayCastVisualizations;
+		} else if (propName == "DrawPixelCheckVisualizations") {
+			reader >> g_SceneMan.m_DrawPixelCheckVisualizations;
+		} else if (propName == "PrintDebugInfo") {
+			reader >> m_PrintDebugInfo;
+		} else if (propName == "MeasureModuleLoadTime") {
+			reader >> m_MeasureModuleLoadTime;
+		} else if (propName == "PlayerNetworkName") {
+			reader >> m_PlayerNetworkName;
+		} else if (propName == "NetworkServerName") {
+			reader >> m_NetworkServerAddress;
+		} else if (propName == "UseNATService") {
+			reader >> g_NetworkServer.m_UseNATService;
+		} else if (propName == "NATServiceAddress") {
+			reader >> m_NATServiceAddress;
+		} else if (propName == "NATServerName") {
+			reader >> m_NATServerName;
+		} else if (propName == "NATServerPassword") {
+			reader >> m_NATServerPassword;
+		} else if (propName == "UseExperimentalMultiplayerSpeedBoosts") {
+			reader >> m_UseExperimentalMultiplayerSpeedBoosts;
+		} else if (propName == "ClientInputFps") {
+			reader >> g_NetworkClient.m_ClientInputFps;
+		} else if (propName == "ServerTransmitAsBoxes") {
+			reader >> g_NetworkServer.m_TransmitAsBoxes;
+		} else if (propName == "ServerBoxWidth") {
+			reader >> g_NetworkServer.m_BoxWidth;
+		} else if (propName == "ServerBoxHeight") {
+			reader >> g_NetworkServer.m_BoxHeight;
+		} else if (propName == "ServerUseHighCompression") {
+			reader >> g_NetworkServer.m_UseHighCompression;
+		} else if (propName == "ServerUseFastCompression") {
+			reader >> g_NetworkServer.m_UseFastCompression;
+		} else if (propName == "ServerUseDeltaCompression") {
+			reader >> g_NetworkServer.m_UseDeltaCompression;
+		} else if (propName == "ServerHighCompressionLevel") {
+			reader >> g_NetworkServer.m_HighCompressionLevel;
+		} else if (propName == "ServerFastAccelerationFactor") {
+			reader >> g_NetworkServer.m_FastAccelerationFactor;
+		} else if (propName == "ServerUseInterlacing") {
+			reader >> g_NetworkServer.m_UseInterlacing;
+		} else if (propName == "ServerEncodingFps") {
+			reader >> g_NetworkServer.m_EncodingFps;
+		} else if (propName == "ServerSleepWhenIdle") {
+			reader >> g_NetworkServer.m_SleepWhenIdle;
+		} else if (propName == "ServerSimSleepWhenIdle") {
+			reader >> g_NetworkServer.m_SimSleepWhenIdle;
+		} else if (propName == "VisibleAssemblyGroup") {
+			m_VisibleAssemblyGroupsList.push_back(reader.ReadPropValue());
+		} else if (propName == "DisableMod") {
+			m_DisabledMods.try_emplace(reader.ReadPropValue(), true);
+		} else if (propName == "EnableGlobalScript") {
+			m_EnabledGlobalScripts.try_emplace(reader.ReadPropValue(), true);
+		} else if (propName == "MouseSensitivity") {
+			reader >> g_UInputMan.m_MouseSensitivity;
+		} else if (propName == "Player1Scheme" || propName == "Player2Scheme" || propName == "Player3Scheme" || propName == "Player4Scheme") {
 			for (int player = Players::PlayerOne; player < Players::MaxPlayerCount; player++) {
 				std::string playerNum = std::to_string(player + 1);
 				if (propName == "Player" + playerNum + "Scheme") {
@@ -309,7 +377,6 @@ namespace RTE {
 		writer.NewLine(false);
 		writer.NewPropertyWithValue("DisableLuaJIT", m_DisableLuaJIT);
 		writer.NewPropertyWithValue("RecommendedMOIDCount", m_RecommendedMOIDCount);
-		writer.NewPropertyWithValue("SimplifiedCollisionDetection", m_SimplifiedCollisionDetection);
 		writer.NewPropertyWithValue("SceneBackgroundAutoScaleMode", m_SceneBackgroundAutoScaleMode);
 		writer.NewPropertyWithValue("DisableFactionBuyMenuThemes", m_DisableFactionBuyMenuThemes);
 		writer.NewPropertyWithValue("DisableFactionBuyMenuThemeCursors", m_DisableFactionBuyMenuThemeCursors);