+++ Managers/SettingsMan.h
@@ -70,12 +70,6 @@ namespace RTE {
 		int RecommendedMOIDCount() const { return m_RecommendedMOIDCount; }
 
 		/// <summary>
-		/// Gets whether simplified collision detection (reduced MOID layer sampling) is enabled.
-		/// </summary>
-		/// <returns>Whether simplified collision detection is enabled or not.</returns>
-		bool SimplifiedCollisionDetection() const { return m_SimplifiedCollisionDetection; }
-
-		/// <summary>
 		/// Gets the Scene background layer auto-scaling mode.
 		/// </summary>
 		/// <returns>The Scene background layer auto-scaling mode. 0 for off, 1 for fit screen dimensions and 2 for always upscaled to x2.</returns>
@@ -543,7 +537,6 @@ namespace RTE {
 
 		bool m_DisableLuaJIT; //!< Whether to disable LuaJIT or not. Disabling will skip loading the JIT library entirely as just setting 'jit.off()' seems to have no visible effect.
 		int m_RecommendedMOIDCount; //!< Recommended max MOID's before removing actors from scenes.
-		bool m_SimplifiedCollisionDetection; //!< Whether simplified collision detection (reduced MOID layer sampling) is enabled.
 		int m_SceneBackgroundAutoScaleMode; //!< Scene background layer auto-scaling mode. 0 for off, 1 for fit screen dimensions and 2 for always upscaled to x2.
 		bool m_DisableFactionBuyMenuThemes; //!< Whether faction BuyMenu theme support is disabled.
 		bool m_DisableFactionBuyMenuThemeCursors; //!< Whether custom cursor support in faction BuyMenu themes is disabled.